# src/Rvi.OrderApi.Model/Models/ErrorResponseV1.cs
/*
 * Dutch Insurance Order API
 *
 * API for placing and fetching orders in the Dutch insurance market
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using Rvi.OrderAPI.Model.Converters;

        namespace Rvi.OrderAPI.Model.Models
        { 
            /// <summary>
                /// 
                /// </summary>
            [DataContract]
            public class ErrorResponseV1 : IEquatable<ErrorResponseV1>
            {
                /// <summary>
                    /// Gets or Sets Error
                    /// </summary>
                    [Required]
                [DataMember(Name="error")]
                    public ErrorResponseV1Error Error { get => ErrorValue; set { ErrorValue = value; ErrorIsSet = true; } }
                    private ErrorResponseV1Error ErrorValue;
                    public bool ErrorIsSet { get; set; }

                /// <summary>
                    /// Gets or Sets ValidationError
                    /// </summary>
                [DataMember(Name="validationError")]
                    public List<ValidationErrorResponseV1> ValidationError { get => ValidationErrorValue; set { ValidationErrorValue = value; ValidationErrorIsSet = true; } }
                    private List<ValidationErrorResponseV1> ValidationErrorValue;
                    public bool ValidationErrorIsSet { get; set; }

            /// <summary>
                /// Returns the string presentation of the object
                /// </summary>
            /// <returns>String presentation of the object</returns>
            public override string ToString()
            {
            var sb = new StringBuilder();
            sb.Append("class ErrorResponseV1 {\n");
                sb.Append("  Error: ").Append(Error).Append("\n");
                sb.Append("  ValidationError: ").Append(ValidationError).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
            }

            /// <summary>
                /// Returns the JSON string presentation of the object
                /// </summary>
            /// <returns>JSON string presentation of the object</returns>
            public string ToJson()
            {
            return Newtonsoft.Json.JsonConvert.SerializeObject(this, Newtonsoft.Json.Formatting.Indented);
            }

            /// <summary>
                /// Returns true if objects are equal
                /// </summary>
            /// <param name="obj">Object to be compared</param>
            /// <returns>Boolean</returns>
            public override bool Equals(object obj)
            {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((ErrorResponseV1)obj);
            }

            /// <summary>
                /// Returns true if ErrorResponseV1 instances are equal
                /// </summary>
            /// <param name="other">Instance of ErrorResponseV1 to be compared</param>
            /// <returns>Boolean</returns>
            public bool Equals(ErrorResponseV1 other)
            {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                Error == other.Error ||
                Error != null &&
                Error.Equals(other.Error)
                ) && 
                (
                ValidationError == other.ValidationError ||
                ValidationError != null &&
                other.ValidationError != null &&
                ValidationError.SequenceEqual(other.ValidationError)
                );
            }

            /// <summary>
                /// Gets the hash code
                /// </summary>
            /// <returns>Hash code</returns>
            public override int GetHashCode()
            {
            unchecked // Overflow is fine, just wrap
            {
            var hashCode = 41;
            // Suitable nullity checks etc, of course :)
                if (Error != null)
                hashCode = hashCode * 59 + Error.GetHashCode();
                if (ValidationError != null)
                hashCode = hashCode * 59 + ValidationError.GetHashCode();
            return hashCode;
            }
            }

            #region Operators
            #pragma warning disable 1591

            public static bool operator ==(ErrorResponseV1 left, ErrorResponseV1 right)
            {
            return Equals(left, right);
            }

            public static bool operator !=(ErrorResponseV1 left, ErrorResponseV1 right)
            {
            return !Equals(left, right);
            }

            #pragma warning restore 1591
            #endregion Operators
            }
}