# src/Rvi.OrderApi.Model/Models/ValidationErrorResponseV1.cs
/*
 * Dutch Insurance Order API
 *
 * API for placing and fetching orders in the Dutch insurance market
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using Rvi.OrderAPI.Model.Converters;

        namespace Rvi.OrderAPI.Model.Models
        { 
            /// <summary>
                /// 
                /// </summary>
            [DataContract]
            public class ValidationErrorResponseV1 : IEquatable<ValidationErrorResponseV1>
            {
                /// <summary>
                    /// Gets or Sets Property
                    /// </summary>
                [DataMember(Name="property")]
                    public string Property { get => PropertyValue; set { PropertyValue = value; PropertyIsSet = true; } }
                    private string PropertyValue;
                    public bool PropertyIsSet { get; set; }

                /// <summary>
                    /// Gets or Sets Code
                    /// </summary>
                [DataMember(Name="code")]
                    public string Code { get => CodeValue; set { CodeValue = value; CodeIsSet = true; } }
                    private string CodeValue;
                    public bool CodeIsSet { get; set; }

                /// <summary>
                    /// Gets or Sets Description
                    /// </summary>
                    [MaxLength(255)]
                [DataMember(Name="description")]
                    public string Description { get => DescriptionValue; set { DescriptionValue = value; DescriptionIsSet = true; } }
                    private string DescriptionValue;
                    public bool DescriptionIsSet { get; set; }

            /// <summary>
                /// Returns the string presentation of the object
                /// </summary>
            /// <returns>String presentation of the object</returns>
            public override string ToString()
            {
            var sb = new StringBuilder();
            sb.Append("class ValidationErrorResponseV1 {\n");
                sb.Append("  Property: ").Append(Property).Append("\n");
                sb.Append("  Code: ").Append(Code).Append("\n");
                sb.Append("  Description: ").Append(Description).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
            }

            /// <summary>
                /// Returns the JSON string presentation of the object
                /// </summary>
            /// <returns>JSON string presentation of the object</returns>
            public string ToJson()
            {
            return Newtonsoft.Json.JsonConvert.SerializeObject(this, Newtonsoft.Json.Formatting.Indented);
            }

            /// <summary>
                /// Returns true if objects are equal
                /// </summary>
            /// <param name="obj">Object to be compared</param>
            /// <returns>Boolean</returns>
            public override bool Equals(object obj)
            {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((ValidationErrorResponseV1)obj);
            }

            /// <summary>
                /// Returns true if ValidationErrorResponseV1 instances are equal
                /// </summary>
            /// <param name="other">Instance of ValidationErrorResponseV1 to be compared</param>
            /// <returns>Boolean</returns>
            public bool Equals(ValidationErrorResponseV1 other)
            {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                Property == other.Property ||
                Property != null &&
                Property.Equals(other.Property)
                ) && 
                (
                Code == other.Code ||
                Code != null &&
                Code.Equals(other.Code)
                ) && 
                (
                Description == other.Description ||
                Description != null &&
                Description.Equals(other.Description)
                );
            }

            /// <summary>
                /// Gets the hash code
                /// </summary>
            /// <returns>Hash code</returns>
            public override int GetHashCode()
            {
            unchecked // Overflow is fine, just wrap
            {
            var hashCode = 41;
            // Suitable nullity checks etc, of course :)
                if (Property != null)
                hashCode = hashCode * 59 + Property.GetHashCode();
                if (Code != null)
                hashCode = hashCode * 59 + Code.GetHashCode();
                if (Description != null)
                hashCode = hashCode * 59 + Description.GetHashCode();
            return hashCode;
            }
            }

            #region Operators
            #pragma warning disable 1591

            public static bool operator ==(ValidationErrorResponseV1 left, ValidationErrorResponseV1 right)
            {
            return Equals(left, right);
            }

            public static bool operator !=(ValidationErrorResponseV1 left, ValidationErrorResponseV1 right)
            {
            return !Equals(left, right);
            }

            #pragma warning restore 1591
            #endregion Operators
            }
}